# portal/src/static/openapi.yml
openapi: 3.0.3
info:
  title: Healthcare Services API
  description: This is our Healthcare Services API that let you develop healthcare applications effortlessly.
  version: "1.0"
servers:
  - url: http://localhost:8002
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/doctors:
    get:
      tags:
        - "Doctor Management"
      description: Get all Doctors records
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                [{"id": "665187a854a5210f0ce74683",
                  "name": "Doctor 1",
                  "contact_no": "0123456789"},
                  {"id": "665187c754a5210f0ce74685",
                  "name": "Doctor 2",
                  "contact_no": "9876543210"}]
  
    post:
      tags:
        - "Doctor Management"
      description: Add Doctor Record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: <doctor's name here>
                dob:
                  type: string
                  example: <dob here>
                gender:
                  type: string
                  enum:
                    - MALE
                    - FEMALE
                contact_no:
                  type: string
                  example: <contact_no>
                job_title:
                  type: string
                  example: <job_title>
                speciality:
                  type: string
                  example: <speciality>
                qualification:
                  type: string
                  example: <qualification>
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Doctor has been Added Successfully!
                  patient_name:
                    type: string
                    example: Doctor 1
  /api/doctors/{doctor_id}:
    get:
      tags:
        - "Doctor Management" 
      description: Get a Specific Doctor record
      parameters:
        - name: doctor_id
          in: path
          required: true
          schema:
            type: string
            example: 665187a854a5210f0ce74683
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                {"_id": "665187a854a5210f0ce74683",
                  "name": "doctor1",
                  "age": "38",
                  "contact_no": "0123456789",
                  "dob": "1980-06-15 00:00:00",
                  "gender": "MALE",
                  "job_title": "Cardiologist",
                  "speciality": "Cardiology",
                  "qualification": "MBBBS",
                  "uid": "6651862de66ba56c4dd11a9d"
                  }

    put:
      tags:
        - "Doctor Management"
      description: Update a specific Doctor record
      parameters:
        - name: doctor_id
          in: path
          required: true
          schema:
            type: string
            example: 665187a854a5210f0ce74683
      requestBody:
        required: True
        content:
          application/json:
            schema:
              type: object
              properties:
                field_you_want_to_update:
                  type: string
                  example: updated value
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Record updated Successfully!
                  doctor_id:
                    type: string
                    example: 665187a854a5210f0ce74683

    delete:
      tags:
        - "Doctor Management"
      description: Delete a specific Doctor record
      parameters:
        - name: doctor_id
          in: path
          required: true
          schema:
            type: string
            example: 665187a854a5210f0ce74683
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                message: Record Deleted Successfully!  
...
  /api/appointments:
    get:
      tags:
        - "Appointment Management"
      description: Get all Appointments
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            example: 1
        - name: patient_id
          in: query
          required: false
          schema:
            type: string
            example: 665186afb873a548422f9b79
        - name: doctor_id
          in: query
          required: false
          schema:
            type: string
            example: 665187c754a5210f0ce74685
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                {"Appointments": [{"ID": "6652f7334d97c23168014069",
                                  "Doctor ID": "665187c754a5210f0ce74685",
                                  "Patient ID": "665186afb873a548422f9b79",
                                  "Date": "Fri, 24 May 2024 00:00:00 GMT",
                                  "Status": "cancelled"},
                                  {"ID": "6652f7374d97c2316801406a",
                                  "Doctor ID": "665187c754a5210f0ce74685",
                                  "Patient ID": "665186afb873a548422f9b79",
                                  "Date": "Sat, 25 May 2024 00:00:00 GMT",
                                  "Status": "cancelled"}]}

    post:
      tags:
        - "Appointment Management"
      description: Add Appointment data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                patient_id:
                  type: string
                  example: <patient_id here>
                doctor_id:
                  type: string
                  example: <doctor_id here>
                date:
                  type: string
                  example: <date here>
                status:
                  type: string
                  enum:
                    - pending
                    - done
                    - cancelled
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                message: "Appointment added Successfully!"
  /api/appointments/{app_id}:
    get:
      tags:
        - "Appointment Management"
      description: Get a Specific Appointment data
      parameters:
        - name: app_id
          in: path
          required: true
          schema:
            type: string
            example: 6652f7334d97c23168014069
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                {"Appointment": {"id": 6652f7334d97c23168014069,
                                "Patient ID": "665186afb873a548422f9b79",
                                "Doctor ID": "665187c754a5210f0ce74685",
                                "Date": "2024-05-24T00:00:00.000+00:00",
                                "status": "cancelled"}}

    put:
      tags:
        - "Appointment Management"
      description: Update Appointment Status
      parameters:
        - name: app_id
          in: path
          required: true
          schema:
            type: string
            example: 6652f7334d97c23168014069
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum:
              - done
              - pending
              - cancelled
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                message: Successfully updated appointment status

    delete:
      tags:
        - "Appointment Management"
      description: Delete Appointment
      parameters:
        - name: app_id
          in: path
          required: true
          schema:
            type: string
            example: 6652f7334d97c23168014069
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                message: Record deleted Successfully!
...